A Partial has constraints
